+++ lib/nanoc/toolbox/helpers.rb
@@ -1,4 +1,5 @@
 require 'nanoc/toolbox/helpers/navigation'
+require 'nanoc/toolbox/helpers/gravatar'
 
 module Nanoc::Toolbox::Helpers